Rewrite htif_isasim_t to match our implementation. It now uses htif_riscy_t as the host HTIF. So changing htif_riscy_t incorrectly may crash spike. spike is now single-threaded.
HTIF is ticked after every step of simulation to make sure all tohost
messages are handled immediately.

mtohost is now a FIFO, and always returns 0 when being read. Writing 0
to mtohost will be ignored.

mfromhost is a reg + FIFO, and reg is only written when it is not 0.

Tested with Linux and SPEC.
(465.tonto needs sys_faccessat, but that is disabled in fesvr)

I don't have good way of incorporating htif_riscy.* into this repo. I
just creat symbolic links for them.